requirements.txt
accelerate>=1.1.1
diffusers>=0.33.1
einops>=0.8.0
huggingface-hub>=0.30.2
imageio>=2.36.0
imageio-ffmpeg>=0.5.1
onnxruntime>=1.14.0
opencv-python>=4.10.0.84
peft>=0.15.2
pillow>=10.2.0
safetensors>=0.4.5
tqdm>=4.67.0
transformers>=4.46.3
scikit-image>=0.25.2
modelscope>=1.21.0
torchvision>=0.18.0
gradio>=5.6.0